+++ addOns/openapi/src/main/java/org/zaproxy/zap/extension/openapi/HistoryPersister.java
@@ -19,7 +19,6 @@
  */
 package org.zaproxy.zap.extension.openapi;
 
-import java.awt.EventQueue;
 import org.apache.log4j.Logger;
 import org.parosproxy.paros.control.Control;
 import org.parosproxy.paros.extension.history.ExtensionHistory;
@@ -28,15 +27,22 @@ import org.parosproxy.paros.model.Model;
 import org.parosproxy.paros.network.HttpMessage;
 import org.parosproxy.paros.network.HttpSender;
 import org.zaproxy.zap.extension.openapi.network.RequesterListener;
+import org.zaproxy.zap.utils.ThreadUtils;
 
 public class HistoryPersister implements RequesterListener {
 
-    private static final Logger LOG = Logger.getLogger(ExtensionOpenApi.class);
+    private static final Logger LOG = Logger.getLogger(HistoryPersister.class);
+
+    private final ExtensionHistory extHistory;
+
+    public HistoryPersister() {
+        this.extHistory =
+                Control.getSingleton().getExtensionLoader().getExtension(ExtensionHistory.class);
+    }
 
     @Override
     public void handleMessage(final HttpMessage message, int initiator) {
-        // Add the message to the history panel and sites tree
-        final HistoryReference historyRef;
+        HistoryReference historyRef;
 
         try {
             historyRef =
@@ -47,28 +53,16 @@ public class HistoryPersister implements RequesterListener {
                                     : HistoryReference.TYPE_ZAP_USER,
                             message);
         } catch (Exception e) {
-            LOG.error(e.getMessage(), e);
+            LOG.warn("Failed to persist the message: " + e.getMessage(), e);
             return;
         }
 
-        final ExtensionHistory extHistory =
-                (ExtensionHistory)
-                        Control.getSingleton()
-                                .getExtensionLoader()
-                                .getExtension(ExtensionHistory.NAME);
-        if (extHistory != null) {
-            EventQueue.invokeLater(
-                    new Runnable() {
-
-                        @Override
-                        public void run() {
-                            extHistory.addHistory(historyRef);
-                            Model.getSingleton()
-                                    .getSession()
-                                    .getSiteTree()
-                                    .addPath(historyRef, message);
-                        }
-                    });
-        }
+        ThreadUtils.invokeAndWaitHandled(
+                () -> {
+                    if (extHistory != null) {
+                        extHistory.addHistory(historyRef);
+                    }
+                    Model.getSingleton().getSession().getSiteTree().addPath(historyRef, message);
+                });
     }
 }